Add request service
