+++ src/main/webapp/app/project-select/project-select.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, Output, EventEmitter } from '@angular/core';
 import { Project, SortDirection, SortBy, ProjectComparators } from '../../model/project';
-import { request, fakeProjects, defaultColors } from '../../utils';
+import { defaultColors } from '../../utils';
 import { faArrowDown, faArrowUp, faCircle, faFilter } from '@fortawesome/free-solid-svg-icons';
 import { trigger, state, style, transition, animate } from '@angular/animations';
 import { HttpService } from '../http.service';
@@ -141,8 +141,7 @@ export class ProjectSelectComponent implements OnInit {
   }
 
   ngOnInit() {
-    fakeProjects();
-    this.httpService.request('/list-project-summaries').then(projects => {
+    this.httpService.listProjects().then(projects => {
       // Sort by the selected fields
       projects.sort(ProjectComparators.getComparator(this.currentSortDirection, this.currentSortField));
       // Assign colors based on initial IAM Bindings order
@@ -153,17 +152,5 @@ export class ProjectSelectComponent implements OnInit {
         this.toggleProject(projects[0]);
       }
     });
-    /* request('/list-project-summaries', 'GET').then(r => r.json()).then(projects => {
-      // Sort by the selected fields
-      projects.sort(ProjectComparators.getComparator(this.currentSortDirection, this.currentSortField));
-      // Assign colors based on initial IAM Bindings order
-      projects.forEach((project, index) => project.color = defaultColors[index % defaultColors.length])
-      this.projects = projects;
-      // Add the highest IAM Bindings project by default
-      if (projects.length > 0) {
-        this.toggleProject(projects[0]);
-      }
-    }); */
   }
-
 }
