+++ src/main/webapp/utils.ts
@@ -23,25 +23,6 @@ import { SimpleChange } from '@angular/core';
 export var defaultIsTest: boolean = true;
 export var defaultColors: string[] = ['#3c78d8', '#cc0000', '#ff9900', '#b6d7a8', '#9c27b0'];
 
-/** Used internally to dish out fake responses when requested. Effectively a map from URL to response */
-var fakeResponses: { [key: string]: any } = {};
-
-/** Sends the given request to HTTP if isTest is false, otherwise fakes out the request */
-export async function request(url: string, method: string, body = undefined, isTest: boolean = defaultIsTest): Promise<{ json: any }> {
-  if (isTest) {
-    return new Promise(resolve => {
-      let response = getFake(url, method);
-      // When the user calls json() on the promise, 
-      resolve({ json: () => response });
-    });
-  } else {
-    return fetch(url, {
-      method: method,
-      body: JSON.stringify(body)
-    });
-  }
-}
-
 /** From a given SimpleChange, extract the projects that were added or deleted */
 export function getAdditionsDeletions(change: SimpleChange): { added: Project[], removed: Project[] } {
   let out = { added: [], removed: [] };
@@ -250,183 +231,4 @@ export function removeFromGraph(properties: { options: google.visualization.Line
 
   // Force a refresh of the chart
   properties.columns = [].concat(properties.columns);
-}
-
-/** Gets the fake response for the given request */
-function getFake(url: string, method: string): any {
-  return fakeResponses[url];
-}
-
-/** Sets the faked-out test response for the given url. Response should be a JS object that can be stringified */
-export function setResponse(url: string, response: any) {
-  fakeResponses[url] = response;
-}
-
-/** Generate fake data for project 1 */
-function fakeProject1(): void {
-  let projectId = 'project-1';
-  // Fake data for showing the graph
-  let iamBindings: { [key: number]: number } = {
-    [Date.parse('1 Jun 2020 UTC')]: 131,
-    [Date.parse('2 Jun 2020 UTC')]: 56,
-    [Date.parse('3 Jun 2020 UTC')]: 84,
-    [Date.parse('4 Jun 2020 UTC')]: 101,
-    [Date.parse('5 Jun 2020 UTC')]: 100,
-    [Date.parse('6 Jun 2020 UTC')]: 90,
-    [Date.parse('7 Jun 2020 UTC')]: 66,
-    [Date.parse('8 Jun 2020 UTC')]: 136,
-    [Date.parse('9 Jun 2020 UTC')]: 108,
-    [Date.parse('10 Jun 2020 UTC')]: 50,
-    [Date.parse('11 Jun 2020 UTC')]: 92,
-    [Date.parse('12 Jun 2020 UTC')]: 136,
-    [Date.parse('13 Jun 2020 UTC')]: 55,
-    [Date.parse('14 Jun 2020 UTC')]: 148,
-    [Date.parse('15 Jun 2020 UTC')]: 141,
-    [Date.parse('16 Jun 2020 UTC')]: 64,
-    [Date.parse('17 Jun 2020 UTC')]: 102,
-    [Date.parse('18 Jun 2020 UTC')]: 139,
-    [Date.parse('19 Jun 2020 UTC')]: 87,
-    [Date.parse('20 Jun 2020 UTC')]: 57,
-  };
-  let recommendations: { [key: number]: Recommendation } = {
-    [Date.parse('5 Jun 2020 UTC')]: new Recommendation(projectId, 'Rec 1', RecommenderType.IAM_BINDING, Date.parse('5 Jun 2020 UTC')),
-    [Date.parse('9 Jun 2020 UTC')]: new Recommendation(projectId, 'Rec 2', RecommenderType.IAM_BINDING, Date.parse('9 Jun 2020 UTC')),
-    [Date.parse('17 Jun 2020 UTC')]: new Recommendation(projectId, 'Rec 3', RecommenderType.IAM_BINDING, Date.parse('17 Jun 2020 UTC')),
-    // Simulate two recommendations on one day
-    [Date.parse('17 Jun 2020 UTC') + 1]: new Recommendation(projectId, 'Rec 4', RecommenderType.IAM_BINDING, Date.parse('17 Jun 2020 UTC') + 1),
-  }
-
-  let url = `/get-project-data?id="${projectId}"`;
-  // Fake out the given url to the generated fake project
-  setResponse(url, new ProjectGraphData(projectId, iamBindings, recommendations));
-}
-
-/** Generate fake data for project 2 */
-function fakeProject2(): void {
-  let projectId = 'project-2';
-  // Fake data for showing the graph
-  let iamBindings: { [key: number]: number } = {
-    [Date.parse('1 Jun 2020 UTC')]: 28,
-    [Date.parse('2 Jun 2020 UTC')]: 36,
-    [Date.parse('3 Jun 2020 UTC')]: 22,
-    [Date.parse('4 Jun 2020 UTC')]: 62,
-    [Date.parse('5 Jun 2020 UTC')]: 60,
-    [Date.parse('6 Jun 2020 UTC')]: 41,
-    [Date.parse('7 Jun 2020 UTC')]: 52,
-    [Date.parse('8 Jun 2020 UTC')]: 27,
-    [Date.parse('9 Jun 2020 UTC')]: 55,
-    [Date.parse('10 Jun 2020 UTC')]: 38,
-    [Date.parse('11 Jun 2020 UTC')]: 28,
-    [Date.parse('12 Jun 2020 UTC')]: 38,
-    [Date.parse('13 Jun 2020 UTC')]: 34,
-    [Date.parse('14 Jun 2020 UTC')]: 18,
-    [Date.parse('15 Jun 2020 UTC')]: 12,
-    [Date.parse('16 Jun 2020 UTC')]: 48,
-    [Date.parse('17 Jun 2020 UTC')]: 47,
-    [Date.parse('18 Jun 2020 UTC')]: 60,
-    [Date.parse('19 Jun 2020 UTC')]: 20,
-    [Date.parse('20 Jun 2020 UTC')]: 47,
-  };
-  let recommendations: { [key: number]: Recommendation } = {
-    [Date.parse('1 Jun 2020 UTC')]: new Recommendation(projectId, 'Rec 1', RecommenderType.IAM_BINDING, Date.parse('1 Jun 2020 UTC')),
-    [Date.parse('9 Jun 2020 UTC')]: new Recommendation(projectId, 'Rec 2', RecommenderType.IAM_BINDING, Date.parse('9 Jun 2020 UTC')),
-    [Date.parse('20 Jun 2020 UTC')]: new Recommendation(projectId, 'Rec 3', RecommenderType.IAM_BINDING, Date.parse('20 Jun 2020 UTC')),
-    // Simulate two recommendations on one day
-    [Date.parse('20 Jun 2020 UTC') + 1]: new Recommendation(projectId, 'Rec 4', RecommenderType.IAM_BINDING, Date.parse('20 Jun 2020 UTC') + 1),
-  }
-
-  let url = `/get-project-data?id="${projectId}"`;
-  // Fake out the given url to the generated fake project
-  setResponse(url, new ProjectGraphData(projectId, iamBindings, recommendations));
-}
-
-/** Generate fake data for project 3 */
-function fakeProject3(): void {
-  let projectId = 'test-long-project-id-project-3';
-  // Fake data for showing the graph
-  let iamBindings: { [key: number]: number } = {
-    [Date.parse('6 Jun 2020 UTC')]: 125,
-    [Date.parse('7 Jun 2020 UTC')]: 201,
-    [Date.parse('8 Jun 2020 UTC')]: 177,
-    [Date.parse('9 Jun 2020 UTC')]: 111,
-    [Date.parse('10 Jun 2020 UTC')]: 212,
-    [Date.parse('11 Jun 2020 UTC')]: 190,
-    [Date.parse('12 Jun 2020 UTC')]: 184,
-    [Date.parse('13 Jun 2020 UTC')]: 137,
-    [Date.parse('14 Jun 2020 UTC')]: 124,
-    [Date.parse('15 Jun 2020 UTC')]: 205,
-    [Date.parse('16 Jun 2020 UTC')]: 182,
-    [Date.parse('17 Jun 2020 UTC')]: 109,
-    [Date.parse('18 Jun 2020 UTC')]: 191,
-    [Date.parse('19 Jun 2020 UTC')]: 211,
-    [Date.parse('20 Jun 2020 UTC')]: 213,
-    [Date.parse('21 Jun 2020 UTC')]: 177,
-    [Date.parse('22 Jun 2020 UTC')]: 136,
-    [Date.parse('23 Jun 2020 UTC')]: 193,
-    [Date.parse('24 Jun 2020 UTC')]: 153,
-    [Date.parse('25 Jun 2020 UTC')]: 187,
-  };
-  let recommendations: { [key: number]: Recommendation } = {
-    [Date.parse('7 Jun 2020 UTC')]: new Recommendation(projectId, 'Rec 1', RecommenderType.IAM_BINDING, Date.parse('7 Jun 2020 UTC')),
-    [Date.parse('9 Jun 2020 UTC')]: new Recommendation(projectId, 'Rec 2', RecommenderType.IAM_BINDING, Date.parse('9 Jun 2020 UTC')),
-    [Date.parse('22 Jun 2020 UTC')]: new Recommendation(projectId, 'Rec 3', RecommenderType.IAM_BINDING, Date.parse('22 Jun 2020 UTC')),
-    // Simulate two recommendations on one day
-    [Date.parse('122 Jun 2020 UTC') + 1]: new Recommendation(projectId, 'Rec 4', RecommenderType.IAM_BINDING, Date.parse('22 Jun 2020 UTC') + 1),
-  }
-
-  let url = `/get-project-data?id="${projectId}"`;
-  // Fake out the given url to the generated fake project
-  setResponse(url, new ProjectGraphData(projectId, iamBindings, recommendations));
-}
-
-/** Generate fake data for project 4 */
-function fakeProject4(): void {
-  let projectId = 'quite-the-long-project-4';
-  // Fake data for showing the graph
-  let iamBindings: { [key: number]: number } = {
-    [Date.parse('1 Jul 2020 UTC')]: 14,
-    [Date.parse('2 Jul 2020 UTC')]: 24,
-    [Date.parse('3 Jul 2020 UTC')]: 23,
-    [Date.parse('4 Jul 2020 UTC')]: 35,
-    [Date.parse('5 Jul 2020 UTC')]: 38,
-    [Date.parse('6 Jul 2020 UTC')]: 19,
-    [Date.parse('7 Jul 2020 UTC')]: 17,
-    [Date.parse('8 Jul 2020 UTC')]: 21,
-    [Date.parse('9 Jul 2020 UTC')]: 12,
-    [Date.parse('10 Jul 2020 UTC')]: 39,
-    [Date.parse('11 Jul 2020 UTC')]: 35,
-    [Date.parse('12 Jul 2020 UTC')]: 15,
-    [Date.parse('13 Jul 2020 UTC')]: 26,
-    [Date.parse('14 Jul 2020 UTC')]: 40,
-    [Date.parse('15 Jul 2020 UTC')]: 36,
-    [Date.parse('16 Jul 2020 UTC')]: 37,
-    [Date.parse('17 Jul 2020 UTC')]: 26,
-    [Date.parse('18 Jul 2020 UTC')]: 28,
-    [Date.parse('19 Jul 2020 UTC')]: 34,
-    [Date.parse('20 Jul 2020 UTC')]: 27,
-  };
-  let recommendations: { [key: number]: Recommendation } = {
-    [Date.parse('1 Jul 2020 UTC')]: new Recommendation(projectId, 'Rec 1', RecommenderType.IAM_BINDING, Date.parse('1 Jul 2020 UTC')),
-    [Date.parse('9 Jul 2020 UTC')]: new Recommendation(projectId, 'Rec 2', RecommenderType.IAM_BINDING, Date.parse('9 Jul 2020 UTC')),
-    [Date.parse('20 Jul 2020 UTC')]: new Recommendation(projectId, 'Rec 3', RecommenderType.IAM_BINDING, Date.parse('20 Jul 2020 UTC')),
-    // Simulate two recommendations on one day
-    [Date.parse('20 Jul 2020 UTC') + 1]: new Recommendation(projectId, 'Rec 4', RecommenderType.IAM_BINDING, Date.parse('20 Jul 2020 UTC') + 1),
-  }
-
-  let url = `/get-project-data?id="${projectId}"`;
-  // Fake out the given url to the generated fake project
-  setResponse(url, new ProjectGraphData(projectId, iamBindings, recommendations));
-}
-
-/** Generate fake data for projects 1 and 2 and sets the appropriate response from request() */
-export function fakeProjects(): void {
-  let prj1 = new Project('Project 1', 'project-1', 1, new ProjectMetaData(100));
-  let prj2 = new Project('Project 2', 'project-2', 2, new ProjectMetaData(70));
-  let prj3 = new Project('Test Long Project ID Project 3', 'test-long-project-id-project-3', 3, new ProjectMetaData(173));
-  let prj4 = new Project('Quite the Long Project 4', 'quite-the-long-project-4', 4, new ProjectMetaData(33));
-  setResponse('/list-project-summaries', [prj1, prj2, prj3, prj4]);
-  fakeProject1();
-  fakeProject2();
-  fakeProject3();
-  fakeProject4();
 }
